Fuseki 5 & Nextgraph Integration

+++ src/middleware/tests/ldp/binary.test.ts
@@ -1,11 +1,10 @@
 import fetch from 'node-fetch';
-import waitForExpect from 'wait-for-expect';
 import fs from 'fs';
 import { ServiceBroker } from 'moleculer';
 import path, { join as pathJoin } from 'path';
-import { getSlugFromUri } from '@semapps/ldp';
+import { Binary } from '@semapps/ldp';
 import { fileURLToPath } from 'url';
-import { fetchServer, createAccount } from '../utils.ts';
+import { createAccount, clearAllDatasets, backupAllDatasets } from '../utils.ts';
 import initialize from './initialize.ts';
 import * as CONFIG from '../config.ts';
 
@@ -15,65 +14,84 @@ jest.setTimeout(20000);
 let broker: ServiceBroker;
 let alice: any;
 
-beforeAll(async () => {
-  broker = await initialize(false);
-  await broker.start();
-  alice = await createAccount(broker, 'alice');
-});
+describe.each(['ng', 'fuseki'])('Binary handling with triplestore %s', (triplestore: string) => {
+  beforeAll(async () => {
+    broker = await initialize(triplestore);
+    await broker.start();
+    await clearAllDatasets(broker);
+    alice = await createAccount(broker, 'alice7');
+  });
 
-afterAll(async () => {
-  if (broker) await broker.stop();
-});
+  afterAll(async () => {
+    if (broker) {
+      if (triplestore === 'ng') await backupAllDatasets(broker); // Allow to see what was persisted
+      await broker.stop();
+    }
+  });
 
-describe('Binary handling of LDP server', () => {
-  let fileUri: string | null;
+  let fileUri: string;
   let filePath: string | null;
-  let fileName: string | null;
+  let binary: Binary;
   let containerUri: string;
 
   test('Post image to container', async () => {
-    // @ts-expect-error This expression is not callable
-    await waitForExpect(async () => {
-      containerUri = await alice.call('ldp.registry.getUri', { type: 'semapps:File', isContainer: true });
-      expect(containerUri).not.toBeUndefined();
-    });
+    containerUri = await alice.getContainerUri(
+      'https://www.w3.org/ns/iana/media-types/application/octet-stream#Resource'
+    );
 
     const readStream = fs.createReadStream(pathJoin(__dirname, 'av-icon.png'));
 
-    const { headers } = await fetchServer(containerUri, {
+    const { headers } = await alice.fetch(containerUri, {
       method: 'POST',
       body: readStream,
       headers: new fetch.Headers({ 'Content-Type': 'image/png' })
     });
 
-    fileUri = headers.get('Location');
+    fileUri = headers.get('Location')!;
     expect(fileUri).not.toBeNull();
 
     filePath = fileUri!.replace(CONFIG.HOME_URL!, '');
-    fileName = getSlugFromUri(fileUri);
 
-    expect(fs.existsSync(pathJoin(__dirname, '../uploads', filePath))).toBeTruthy();
+    if (triplestore === 'fuseki') {
+      expect(fs.existsSync(pathJoin(__dirname, '../uploads', filePath))).toBeTruthy();
+    }
   });
 
-  test('Get container', async () => {
-    await expect(fetchServer(containerUri)).resolves.toMatchObject({
-      json: {
-        type: expect.arrayContaining(['ldp:Container', 'ldp:BasicContainer']),
-        'ldp:contains': [
-          {
-            id: fileUri,
-            type: 'semapps:File',
-            'semapps:fileName': fileName,
-            'semapps:localPath': `uploads/${filePath}`,
-            'semapps:mimeType': 'image/png'
-          }
-        ]
-      }
+  test('Get binary (via Moleculer action)', async () => {
+    binary = await alice.call('ldp.binary.get', { resourceUri: fileUri });
+
+    expect(binary).toMatchObject({
+      file: expect.anything(),
+      mimeType: 'image/png',
+      size: 3181,
+      time: triplestore === 'fuseki' ? expect.anything() : undefined
     });
   });
 
+  test('Get container (via API)', async () => {
+    const { json } = await alice.fetch(containerUri);
+
+    expect(json).toMatchObject({
+      type: expect.arrayContaining(['ldp:Container', 'ldp:BasicContainer']),
+      'ldp:contains': [
+        {
+          id: fileUri,
+          type: expect.arrayContaining([
+            'https://www.w3.org/ns/iana/media-types/application/octet-stream#Resource',
+            'https://www.w3.org/ns/iana/media-types/image/png#Resource'
+          ]),
+          'stat:size': '3181'
+        }
+      ]
+    });
+
+    if (triplestore === 'fuseki') {
+      expect(json['ldp:contains'][0]['stat:mtime']).toBe(binary.time?.toISOString());
+    }
+  });
+
   test('Get image as binary (via API)', async () => {
-    const { headers, body } = await fetchServer(fileUri, {
+    const { headers, body } = await alice.fetch(fileUri, {
       headers: new fetch.Headers({ Accept: '*/*' })
     });
 
@@ -84,34 +102,62 @@ describe('Binary handling of LDP server', () => {
     expect(body).toContain('PNG');
   });
 
-  test('Get image as resource (via Moleculer action)', async () => {
-    await expect(alice.call('ldp.resource.get', { resourceUri: fileUri })).resolves.toMatchObject({
+  test('Get image as resource (via API)', async () => {
+    const { json } = await alice.fetch(fileUri, {
+      headers: new fetch.Headers({ Accept: 'application/ld+json' })
+    });
+
+    expect(json).toMatchObject({
       id: fileUri,
-      type: 'semapps:File',
-      'semapps:fileName': fileName,
-      'semapps:localPath': `uploads/${filePath}`,
-      'semapps:mimeType': 'image/png'
+      type: expect.arrayContaining([
+        'https://www.w3.org/ns/iana/media-types/application/octet-stream#Resource',
+        'https://www.w3.org/ns/iana/media-types/image/png#Resource'
+      ]),
+      'stat:size': '3181'
     });
+
+    if (triplestore === 'fuseki') {
+      expect(json['stat:mtime']).toBe(binary.time?.toISOString());
+    }
   });
 
-  test('Delete image', async () => {
-    await expect(
-      fetchServer(fileUri, {
-        method: 'DELETE'
-      })
-    ).resolves.toMatchObject({
-      status: 204
+  test('Get image as resource (via Moleculer action)', async () => {
+    const resource = await alice.call('ldp.resource.get', { resourceUri: fileUri });
+
+    expect(resource).toMatchObject({
+      id: fileUri,
+      type: expect.arrayContaining([
+        'https://www.w3.org/ns/iana/media-types/application/octet-stream#Resource',
+        'https://www.w3.org/ns/iana/media-types/image/png#Resource'
+      ]),
+      'stat:size': '3181'
     });
 
-    expect(fs.existsSync(pathJoin(__dirname, '../uploads/files/av-icon.png'))).toBeFalsy();
-
-    await expect(fetchServer(fileUri)).resolves.toMatchObject({ status: 404 });
-
-    await expect(fetchServer(containerUri)).resolves.toMatchObject({
-      json: {
-        type: expect.arrayContaining(['ldp:Container', 'ldp:BasicContainer']),
-        'ldp:contains': []
-      }
-    });
+    if (triplestore === 'fuseki') {
+      expect(resource['stat:mtime']).toBe(binary.time?.toISOString());
+    }
   });
+
+  if (triplestore === 'fuseki') {
+    test('Delete image (via API)', async () => {
+      await expect(
+        alice.fetch(fileUri, {
+          method: 'DELETE'
+        })
+      ).resolves.toMatchObject({
+        status: 204
+      });
+
+      expect(fs.existsSync(pathJoin(__dirname, '../uploads/files/av-icon.png'))).toBeFalsy();
+
+      await expect(alice.fetch(fileUri)).resolves.toMatchObject({ status: 404 });
+
+      await expect(alice.fetch(containerUri)).resolves.toMatchObject({
+        json: {
+          type: expect.arrayContaining(['ldp:Container', 'ldp:BasicContainer']),
+          'ldp:contains': []
+        }
+      });
+    });
+  }
 });
